Renamed spider package
