+++ review/pkg/crawler/go.mod
@@ -1,4 +1,4 @@
-module spider
+module crawler
 
 go 1.15
