+++ review/pkg/engine/go.mod
@@ -1,9 +1,3 @@
 module engine
 
 go 1.15
-
-require (
-	golang.org/x/net v0.0.0-20201031054903-ff519b6c9102 // indirect
-	goondex/goondex v0.0.0
-	goondex/spider v0.0.0
-)
